fix stack trace alignment & removed in app badge
