+++ static/app/components/events/interfaces/frame/lineV2/native.tsx
@@ -19,13 +19,13 @@ import {getPlatform} from '../utils';
 import Expander from './expander';
 import GroupingBadges from './groupingBadges';
 import LeadHint from './leadHint';
+import Wrapper from './wrapper';
 
 type Props = React.ComponentProps<typeof Expander> &
   React.ComponentProps<typeof LeadHint> &
   Omit<React.ComponentProps<typeof GroupingBadges>, 'inApp'> & {
     frame: Frame;
     leadsToApp: boolean;
-    hasGroupingBadge: boolean;
     isFrameAfterLastNonApp?: boolean;
     includeSystemFrames?: boolean;
     showingAbsoluteAddress?: boolean;
@@ -46,7 +46,6 @@ function Native({
   isExpanded,
   nextFrame,
   leadsToApp,
-  hasGroupingBadge,
   isHoverPreviewed,
   onAddressToggle,
   image,
@@ -64,7 +63,7 @@ function Native({
   haveFramesAtLeastOneGroupingBadge,
   ...props
 }: Props) {
-  const {instructionAddr, trust, addrMode, inApp, symbolicatorStatus} = frame ?? {};
+  const {instructionAddr, trust, addrMode, symbolicatorStatus} = frame ?? {};
 
   function packageStatus() {
     // this is the status of image that belongs to this frame
@@ -161,9 +160,8 @@ function Native({
           isHoverPreviewed={isHoverPreviewed}
         />
       </NativeLineContent>
-      {hasGroupingBadge && (
+      {haveFramesAtLeastOneGroupingBadge && (
         <GroupingBadges
-          inApp={inApp}
           isPrefix={isPrefix}
           isSentinel={isSentinel}
           isUsedForGrouping={isUsedForGrouping}
@@ -181,33 +179,11 @@ function Native({
 
 export default Native;
 
-const Wrapper = styled('div')<{
-  haveFramesAtLeastOneExpandedFrame?: boolean;
-  haveFramesAtLeastOneGroupingBadge?: boolean;
-}>`
-  display: grid;
-  grid-template-columns: ${p =>
-    p.haveFramesAtLeastOneGroupingBadge && p.haveFramesAtLeastOneExpandedFrame
-      ? '1.5fr 0.5fr 16px'
-      : p.haveFramesAtLeastOneGroupingBadge
-      ? '1fr 0.5fr'
-      : p.haveFramesAtLeastOneExpandedFrame
-      ? '1fr 16px'
-      : '1fr'};
-
-  grid-gap: ${space(1)};
-
-  @media (min-width: ${props => props.theme.breakpoints[0]}) {
-    align-items: center;
-  }
-`;
-
 const NativeLineContent = styled('div')<{isFrameAfterLastNonApp: boolean}>`
   display: grid;
   flex: 1;
   grid-gap: ${space(0.5)};
-  grid-template-columns: ${p =>
-    `minmax(${p.isFrameAfterLastNonApp ? '167px' : '117px'}, auto)  1fr`};
+  grid-template-columns: auto 1fr;
   align-items: center;
   justify-content: flex-start;
