+++ static/app/types/index.tsx
@@ -1991,7 +1991,6 @@ export enum FrameBadge {
   SENTINEL = 'sentinel',
   PREFIX = 'prefix',
   GROUPING = 'grouping',
-  IN_APP = 'in_app',
 }
 
 /**
